requirements.txt
aiofiles==23.2.1
asgiref==3.7.2
attrs==23.2.0
Django==4.2.6
django-filter==24.2
django-json-widget==1.1.1
django-rest-framework==0.1.0
djangoql==0.1.1
djangorestframework==3.15.1
djangorestframework-api-key==3.0.0
drf-spectacular==0.27.2
future==0.18.3
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==24.0
ply==3.11
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
